+++ lib/src/_index.dart
@@ -1,10 +1,12 @@
 // BSD 3-Clause License
-// Copyright (c) 2022, GM Consult Pty Ltd
+// Copyright ©2022, GM Consult Pty Ltd
 // All rights reserved
 
 export 'package:porter_2_stemmer/porter_2_stemmer.dart';
 export 'package:porter_2_stemmer/extensions.dart';
 export 'english_analyzer/english.dart';
+export 'part_of_speech.dart';
+export 'suggestion.dart';
 export 'term_similarity.dart';
 export 'text_analyzer.dart';
 export 'text_document.dart';