+++ lib/src/part_of_speech_tagging/_index.dart
@@ -1,0 +1,6 @@
+// Copyright ©2022, GM Consult (Pty) Ltd.
+// BSD 3-Clause Licensedictosaurus
+// All rights reserved
+
+export 'part_of_speech.dart';
+export 'pos_tag.dart';