+++ lib/src/text_tokenizer.dart
@@ -1,5 +1,5 @@
 // BSD 3-Clause License
-// Copyright (c) 2022, GM Consult Pty Ltd
+// Copyright ©2022, GM Consult Pty Ltd
 // All rights reserved
 
 import '_index.dart';