+++ lib/type_definitions.dart
@@ -1,5 +1,5 @@
 // BSD 3-Clause License
-// Copyright (c) 2022, GM Consult Pty Ltd
+// Copyright ©2022, GM Consult Pty Ltd
 // All rights reserved
 
 /// Exports all the type definitions used in the `text_analysis` library.